Migrating to a universal code base: created enums.rs and place procees status in the file. move Process type to src/process.rs. Done only for macOS. Will do Linux if no objection of moving ahead with the type of changes.
