+++ src/mac/mod.rs
@@ -7,12 +7,12 @@
 pub mod component;
 pub mod disk;
 mod ffi;
-pub mod process;
 pub mod processor;
 pub mod system;
 
 pub use self::component::Component;
 pub use self::disk::{Disk, DiskType};
-pub use self::process::{Process,ProcessStatus};
 pub use self::processor::Processor;
 pub use self::system::System;
+pub use process::Process;
+pub use enums::ProcessStatus;
